pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "kindle-wikipedia-cli"
version = "0.1.0"
description = "A CLI tool to send Wikipedia articles to Kindle"
authors = [
  { name = "Ishan Rai", email = "ishanrai@example.com" },
]
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "typer[all]",
    "requests",
    "beautifulsoup4",
    "python-dotenv",
    "rich",
]

[project.scripts]
send-to-kindle = "src.cli:app"

[tool.setuptools.packages.find]
where = ["."]
